third_party/blink/web_tests/external/wpt/speculation-rules/prefetch/tentative/cookie-indices.https_include=changedWithRedirect-expected.txt

This is a testharness.js-based test.
[FAIL] Cookie-Indices should prevent a prefetch from succeeding if the cookie changed, with a redirect
  assert_equals: expected (undefined) undefined but got (string) "prefetch"
Harness: the test ran to completion.
